# docs/spec/openapi.yaml
openapi: 3.0.3
info:
  title: STDS API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://example.com/licenses/proprietary
  description: |
    STDS（物業代管系統）API 規格
    產出依據：docs/design/domain-model.md v3.1
    授權模型：混合型（RBAC + Resource-based）
    - x-required-role：列出可呼叫的角色（RBAC）
    - x-ownership-required：物業層級存取控制（resource-based），透過 Firebase ID token 內 Custom Claims 的 assigned_property_ids 驗證
    - 優先順序：先驗 RBAC role，再驗 property ownership（兩者皆須通過）

    認證機制（v3.1）：系統使用 Firebase Auth 管理使用者認證。
    - 登入流程：由前端 Firebase SDK 處理，後端不接受 email/password 登入
    - 後端驗證：每個 request 以 Firebase ID token（Bearer token）呼叫後端
    - 後端 middleware 以 Firebase Admin SDK 驗證 token，從 Custom Claims 讀取 role 與 assigned_property_ids
    - 密碼重設：由 Firebase SDK 處理，後端不提供密碼重設 API
    - users table 透過 firebase_uid 與 Firebase 帳號對應

    ## Notification 機制

    Email 通知由後端 event handler 統一處理，不提供 REST endpoint：
    - RentBillsGenerated → 寄送帳單通知給租客
    - LeaseExpiringSoon → 寄送到期提醒給 organizer/staff
    - OverdueBillReminder → 寄送催收通知（最多 3 次）
    - FinancialReportSend → 寄送財報 Email

    通知服務由後端訂閱 Domain Event 後透過 SendGrid 發送，前端無需呼叫任何 notification API。
servers:
  - url: /api/v1
security:
  - BearerAuth: []
paths:
  /auth/sync:
    post:
      operationId: syncAuth
      summary: 同步 Firebase 使用者至後端 DB（首次登入或 token 更新時呼叫）
      description: |
        前端以 Firebase SDK 完成登入後，帶 Firebase ID token 呼叫此 endpoint。
        後端以 firebase_uid 查詢既有 users table 記錄，並將最新 role 與 assigned_property_ids
        寫入 Firebase Custom Claims（若與 DB 不一致時更新）。
        此 endpoint 取代原本的 POST /auth/login（v2.9），登入由 Firebase SDK 處理。
      tags:
        - Identity & Access
      responses:
        '200':
          description: 同步成功，回傳使用者基本資料與 Custom Claims 資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: 同步成功
                  value:
                    id: 00000000-0000-0000-0000-000000000001
                    firebase_uid: uid_abc123xyz
                    name: 王小明
                    role: organizer
                    assigned_property_ids:
                      - 10000000-0000-0000-0000-000000000001
                      - 10000000-0000-0000-0000-000000000002
        '401':
          description: Firebase ID token 無效或已過期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_token:
                  summary: token 無效
                  value:
                    error_code: INVALID_FIREBASE_TOKEN
                    message: Firebase ID token 無效或已過期
        '404':
          description: 此 firebase_uid 尚未在後端建立帳號
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: 帳號尚未建立
                  value:
                    error_code: USER_NOT_FOUND
                    message: 此 Firebase 帳號尚未在系統中建立，請聯絡管理員
  /users:
    get:
      operationId: listUsers
      summary: 成員列表
      description: |
        查詢工作室成員列表，可依 role 篩選。
        x-required-role：admin 可查全部；業主無權限。
      tags:
        - Identity & Access
      x-required-role:
        - admin
        - organizer
        - staff
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum:
              - admin
              - organizer
              - staff
              - owner
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 成員列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              examples:
                success:
                  summary: 成員列表
                  value:
                    data:
                      - id: 00000000-0000-0000-0000-000000000001
                        name: 王小明
                        email: organizer@studio.com
                        role: organizer
                      - id: 00000000-0000-0000-0000-000000000002
                        name: 李小花
                        email: staff@studio.com
                        role: staff
        '401':
          description: 未認證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  summary: 未提供 token
                  value:
                    error_code: UNAUTHORIZED
                    message: 請先登入
        '403':
          description: 權限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  summary: 業主無法查看成員列表
                  value:
                    error_code: FORBIDDEN
                    message: 權限不足
    post:
      operationId: createUser
      summary: 建立使用者帳號
      description: |
        在後端 DB 建立使用者記錄（對應 Firebase 帳號已在 Firebase Console 或由 Firebase Admin SDK 建立）。
        - 建立工作室成員：x-required-role: admin
        - 建立業主帳號：x-required-role: admin, organizer
        注意：帳號建立後需另行在 Firebase 建立對應帳號（或由 admin SDK 呼叫），密碼由 Firebase 管理，此 API 不接受 password 欄位。
      tags:
        - Identity & Access
      x-required-role:
        - admin
        - organizer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              create_staff:
                summary: 建立員工帳號
                value:
                  firebase_uid: uid_staff_def456
                  email: newstaff@studio.com
                  name: 張大維
                  role: staff
              create_owner:
                summary: 建立業主帳號
                value:
                  firebase_uid: uid_owner_ghi789
                  email: owner@example.com
                  name: 陳老闆
                  role: owner
      responses:
        '201':
          description: 帳號建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: 建立成功
                  value:
                    id: 00000000-0000-0000-0000-000000000003
                    firebase_uid: uid_staff_def456
                    email: newstaff@studio.com
                    name: 張大維
                    role: staff
                    created_at: '2026-04-15T10:00:00Z'
        '400':
          description: 欄位驗證錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_email:
                  summary: email 未填
                  value:
                    error_code: VALIDATION_EMAIL_REQUIRED
                    message: email 為必填欄位
                invalid_email:
                  summary: email 格式錯誤
                  value:
                    error_code: VALIDATION_EMAIL_INVALID
                    message: email 格式不正確
                missing_firebase_uid:
                  summary: firebase_uid 未填
                  value:
                    error_code: VALIDATION_FIREBASE_UID_REQUIRED
                    message: firebase_uid 為必填欄位
        '401':
          description: 未認證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  summary: 未提供 token
                  value:
                    error_code: UNAUTHORIZED
                    message: 請先登入
        '403':
          description: 權限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  summary: 員工無法建立成員帳號
                  value:
                    error_code: FORBIDDEN
                    message: 權限不足
        '409':
          description: Email 或 firebase_uid 已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_exists:
                  summary: Email 重複
                  value:
                    error_code: EMAIL_ALREADY_EXISTS
                    message: 此 Email 已被使用
                firebase_uid_exists:
                  summary: firebase_uid 重複
                  value:
                    error_code: FIREBASE_UID_ALREADY_EXISTS
                    message: 此 Firebase UID 已被使用
  /users/me:
    get:
      operationId: getCurrentUser
      summary: 查看自己的帳號資料
      tags:
        - Identity & Access
      x-required-role:
        - admin
        - organizer
        - staff
        - owner
      responses:
        '200':
          description: 自己的帳號資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: 帳號資料
                  value:
                    id: 00000000-0000-0000-0000-000000000001
                    firebase_uid: uid_abc123xyz
                    name: 王小明
                    email: organizer@studio.com
                    role: organizer
                    assigned_property_ids:
                      - 10000000-0000-0000-0000-000000000001
                      - 10000000-0000-0000-0000-000000000002
        '401':
          description: 未認證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  summary: 未提供 token
                  value:
                    error_code: UNAUTHORIZED
                    message: 請先登入
  /users/{id}:
    get:
      operationId: getUser
      summary: 成員詳情
      tags:
        - Identity & Access
      x-required-role:
        - admin
        - organizer
        - staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成員詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: 成員詳情
                  value:
                    id: 00000000-0000-0000-0000-000000000001
                    firebase_uid: uid_abc123xyz
                    name: 王小明
                    email: organizer@studio.com
                    role: organizer
                    assigned_property_ids:
                      - 10000000-0000-0000-0000-000000000001
                      - 10000000-0000-0000-0000-000000000002
        '401':
          description: 未認證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  value:
                    error_code: UNAUTHORIZED
                    message: 請先登入
        '403':
          description: 權限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  value:
                    error_code: FORBIDDEN
                    message: 權限不足
        '404':
          description: 使用者不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    error_code: USER_NOT_FOUND
                    message: 使用者不存在
    patch:
      operationId: updateUser
      summary: 更新使用者帳號（角色調整）
      tags:
        - Identity & Access
      x-required-role:
        - admin
        - organizer
        - staff
        - owner
      description: |
        - 角色調整：x-required-role: admin
        - 名稱更新：x-required-role: admin, organizer, staff, owner（本人）
        - 密碼重設：由 Firebase SDK 處理，不透過此 API（前端呼叫 Firebase sendPasswordResetEmail）
        - BR-13：系統管理員不得降低自己的角色
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            examples:
              change_role:
                summary: 變更角色
                value:
                  role: organizer
              update_name:
                summary: 更新名稱
                value:
                  name: 王大明
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  value:
                    id: 00000000-0000-0000-0000-000000000001
                    name: 王大明
                    role: organizer
                    updated_at: '2026-04-15T11:00:00Z'
        '401':
          description: 未認證
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unauthorized:
                  value:
                    error_code: UNAUTHORIZED
                    message: 請先登入
        '403':
          description: 權限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  value:
                    error_code: FORBIDDEN
                    message: 權限不足
        '404':
          description: 使用者不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  value:
                    error_code: USER_NOT_FOUND
                    message: 使用者不存在
        '422':
          description: Business Rule 拒絕
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                admin_cannot_downgrade_self:
                  summary: BR-13 系統管理員不得降低自己的角色
                  value:
                    error_code: ADMIN_CANNOT_DOWNGRADE_SELF
                    message: 系統管理員不得降低自己的角色
...
  /force-terminations/{id}:
    get:
      operationId: getForceTermination
      summary: 查詢強制終止進度
      tags:
        - Leasing
      security:
        - BearerAuth: []
      x-required-role:
        - admin
        - organizer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 強制終止進度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForceTerminationResponse'
        '403':
          description: 權限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 找不到強制終止記錄
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase ID token。前端以 Firebase SDK 登入後取得，每個 request 帶入 Authorization header。
        後端以 Firebase Admin SDK 驗證，Custom Claims 包含 role 與 assigned_property_ids。
    SchedulerKey:
      type: apiKey
      in: header
      name: X-Scheduler-Key
      description: |
        外部排程系統呼叫 backend job endpoint 時使用的共享密鑰。
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firebase_uid:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum:
            - admin
            - organizer
            - staff
            - owner
        permission_overrides:
          type: array
          items:
            type: object
        assigned_property_ids:
          type: array
          items:
            type: string
            format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum:
            - admin
            - organizer
            - staff
            - owner
...